requirements.txt
django
django-braces
django-rest-framework
django-webpack
django-webpack-loader
psycopg2
